Implemented result pattern

+++ Ava.Application/Bookings/Commands/DeleteBookingCommand.cs
@@ -1,4 +1,6 @@
-﻿using Ava.Infrastructure.Db;
+﻿using Ava.Domain.Models.Booking;
+using Ava.Domain.Models.Common;
+using Ava.Infrastructure.Db;
 
 using MediatR;
 
@@ -6,9 +8,9 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Ava.Application.Bookings.Commands;
 
-public record DeleteBookingCommand(Guid Id) : IRequest;
+public record DeleteBookingCommand(Guid Id) : IRequest<Result>;
 
-public class DeleteBookingCommandHandler : IRequestHandler<DeleteBookingCommand>
+public class DeleteBookingCommandHandler : IRequestHandler<DeleteBookingCommand, Result>
 {
     private readonly AvaDbContext _context;
     public DeleteBookingCommandHandler(AvaDbContext context)
@@ -16,8 +18,16 @@ public class DeleteBookingCommandHandler : IRequestHandler<DeleteBookingCommand>
         _context = context;
     }
 
-    public async Task Handle(DeleteBookingCommand request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeleteBookingCommand request, CancellationToken cancellationToken)
     {
-        var result = await _context.Bookings.Where(b => b.Id == request.Id).ExecuteDeleteAsync(cancellationToken);
+        var bookingExists = await _context.Bookings.AnyAsync(b => b.Id == request.Id);
+
+        if (!bookingExists)
+        {
+            return Result.Failure(BookingErrors.NotFound);
+        }
+
+        await _context.Bookings.Where(b => b.Id == request.Id).ExecuteDeleteAsync(cancellationToken);
+        return Result.Success();
     }
 }
